requirements.txt
accelerate
bitsandbytes
datasets
#deepspeed==0.15.0
deepspeed
einops
#flash-attn==2.7.0.post2
flash-attn
isort
jsonlines
loralib
optimum
packaging
peft
#ray[default]==2.12.0
ray[default]
tensorboard
#torch
torchmetrics
tqdm
transformers==4.46.3
transformers_stream_generator
wandb
wheel
